Rename Mobile Modal Form Step 2 component
Change Suffix from 'Mobile' to 'MobileModal' to express what this does.

See comment https://github.com/wmde/fundraising-banners/pull/1126#pullrequestreview-1182505885
